WIP rebuild overview by deriving from the core participants table
This is monstrously complicated particularly when dynamic behaviour
comes into play. Presently filters don't work properly because
core_user/participantsfilter is special. That's before I even try
hooking in the bars column. I suspect by the end we will have
subclassed and copied a hell of a lot of the core participants code...